Reconfigure files and folders structures
