+++ package.json
@@ -19,5 +19,12 @@
   "homepage": "https://github.com/damiecode/Tic-Tac-Toe-Js#readme",
   "devDependencies": {
     "jest": "^25.1.0"
+  },
+  "dependencies": {
+    "loadash": "^1.0.0",
+    "resolve-cwd": "^3.0.0",
+    "webpack": "^4.41.6",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   }
 }
