+++ src/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable no-use-before-define */
+/* eslint-env browser */
+
+import * as Game from './js/play';
+
+Game;
