+++ src/js/gameboard.js
@@ -1,0 +1,7 @@
+
+const GameBoard = (() => {
+  const board = ['', '', '', '', '', '', '', '', ''];
+  return { board };
+})();
+
+export default GameBoard;
