+++ src/js/playerFactory.js
@@ -1,0 +1,3 @@
+const PlayerFactory = (name, symbol, score) => ({ name, symbol, score });
+
+export default PlayerFactory;
