+++ test/gameboard.test.js
@@ -1,0 +1,1 @@
+const gameboard = require('./tic-tac-toe');
